fix somw bugs

+++ dgtle.py
@@ -3,8 +3,16 @@ import shutil
 import requests
 from lxml import etree
 
+def parse_name(name): # 处理windows系统下非法的目录字符
+    black_list = ('\\', '/', ':', '*', '?', '"', '<', '>', '|') # 非法字符集
+    name = ' '.join(name.split())
+    name = ''.join(list(map(lambda x: '_' if x in black_list else x, name)))
+    return name
+
 def down_load(url, cnt): # download the target picture
     print('Getting No.%s  %s...' % (cnt, url))
+    if os.path.exists(cnt + '.jpg'): # if exists, do nothing
+        return
     res = requests.get(url, stream = True)
     with open(cnt + '.jpg', 'wb') as fout:
         shutil.copyfileobj(res.raw, fout)
@@ -14,7 +22,7 @@ if __name__ == '__main__':
     url = input('Please Input the URL...\n')
     page = requests.get(url)
     html = etree.HTML(page.text)
-    dir = os.path.join(os.getcwd(), 'dgtle', html.xpath('/html/head/title')[0].text) # get the page title
+    dir = os.path.join(os.getcwd(), 'dgtle', parse_name(html.xpath('/html/head/title')[0].text)) # get the page title
     if not os.path.exists(dir): # creat the directory
         os.makedirs(dir)
     os.chdir(dir)
